segmentron-master/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
backcall==0.1.0
bleach==2.1.3
certifi==2018.4.16
cffi==1.11.5
click==6.7
cloudpickle==0.5.3
constantly==15.1.0
cryptography==2.2.2
cycler==0.10.0
Cython==0.28.4
dask==0.18.2
decorator==4.3.0
entrypoints==0.2.3
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
imageio==2.3.0
incremental==17.5.0
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.3.1
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.6.0
numpy==1.15.0
olefile==0.45.1
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
Pygments==2.2.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
pyzmq==17.0.0
qtconsole==4.3.1
scikit-image==0.14.0
scipy==1.1.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==5.0.2
traitlets==4.3.2
Twisted==18.7.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.3.1
zope.interface==4.5.0
